validator-core - refactor AttributeUniqueValidator and create index for referenced columns in validation Database (refs #224)
